Revert "Session actor perf (#724)"
This reverts commit a196e8b0d1f48975fb37752166de389970b53a42.
