+++ ydb/core/kqp/common/kqp_ru_calc.h
@@ -3,9 +3,6 @@
 #include <util/system/types.h>
 #include <util/datetime/base.h>
 
-#include <ydb/core/protos/kqp_stats.pb.h>
-#include <ydb/core/kqp/executer_actor/kqp_executer_stats.h>
-
 namespace NKqpProto {
     class TKqpStatsQuery;
 }
@@ -17,21 +14,8 @@ struct TProgressStatEntry;
 
 namespace NRuCalc {
 
-ui64 CalcReadIORu(const TTableStat& stat);
-
-class TIoReadStat: public TTableStat {
-public:
-    void Add(const NYql::NDqProto::TDqTableStats& tableAggr);
-    ui64 CalcRu() const;
-};
-
-class TIoWriteStat: public TTableStat {
-public:
-    void Add(const NYql::NDqProto::TDqTableStats& tableAggr);
-    ui64 CalcRu() const;
-};
-
 ui64 CpuTimeToUnit(TDuration cpuTimeUs);
+ui64 CalcRequestUnit(const NKqpProto::TKqpStatsQuery& stats);
 ui64 CalcRequestUnit(const TProgressStatEntry& stats);
 
 } // namespace NRuCalc
