+++ ydb/core/kqp/provider/yql_kikimr_provider.h
@@ -308,44 +308,34 @@ public:
 
         bool hasScheme = false;
         bool hasData = false;
-        for (auto& [_, operation] : TableOperations) {
-            hasScheme = hasScheme || (operation & KikimrSchemeOps());
-            hasData = hasData || (operation & KikimrDataOps());
+        for (auto& pair : TableOperations) {
+            hasScheme = hasScheme || (pair.second & KikimrSchemeOps());
+            hasData = hasData || (pair.second & KikimrDataOps());
         }
 
-        THashMap<TStringBuf, const NKqpProto::TKqpTableInfo*> tableInfoMap;
-        tableInfoMap.reserve(tableInfos.size());
-        if (TableByIdMap.empty()) {
-            TableByIdMap.reserve(tableInfos.size());
-        }
-        if (TableOperations.empty()) {
-            TableOperations.reserve(operations.size());
-        }
-
+        THashMap<TString, NKqpProto::TKqpTableInfo> tableInfoMap;
         for (const auto& info : tableInfos) {
-            tableInfoMap.emplace(info.GetTableName(), &info);
+            tableInfoMap.insert(std::make_pair(info.GetTableName(), info));
 
             TKikimrPathId pathId(info.GetTableId().GetOwnerId(), info.GetTableId().GetTableId());
-            TableByIdMap.emplace(pathId, info.GetTableName());
+            TableByIdMap.insert(std::make_pair(pathId, info.GetTableName()));
         }
 
         for (const auto& op : operations) {
-            const auto& table = [&]() -> const TString& {
-                const auto tempTable = TempTables.FindPtr(op.GetTable());
-                if (tempTable) {
-                    return *tempTable;
-                } else {
-                    return op.GetTable();
-                }
-            }();
+            auto table = op.GetTable();
 
-            const auto newOp = TYdbOperation(op.GetOperation());
+            auto newOp = TYdbOperation(op.GetOperation());
+            TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
+
+            auto tempTable = TempTables.FindPtr(table);
+            if (tempTable) {
+                table = *tempTable;
+            }
 
             const auto info = tableInfoMap.FindPtr(table);
             if (!info) {
                 TString message = TStringBuilder()
                     << "Unable to find table info for table '" << table << "'";
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_SCHEME_ERROR, message));
                 return {false, issues};
             }
@@ -353,7 +343,6 @@ public:
             if (queryType == EKikimrQueryType::Dml && (newOp & KikimrSchemeOps())) {
                 TString message = TStringBuilder() << "Operation '" << newOp
                     << "' can't be performed in data query";
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                 return {false, issues};
             }
@@ -362,7 +351,6 @@ public:
                 if (EffectiveIsolationLevel) {
                     TString message = TStringBuilder() << "Scheme operations can't be performed inside transaction, "
                         << "operation: " << newOp;
-                    const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                     issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                     return {false, issues};
                 }
@@ -371,7 +359,6 @@ public:
             if (queryType == EKikimrQueryType::Ddl && (newOp & KikimrDataOps())) {
                 TString message = TStringBuilder() << "Operation '" << newOp
                     << "' can't be performed in scheme query";
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                 return {false, issues};
             }
@@ -379,7 +366,6 @@ public:
             if (queryType == EKikimrQueryType::Scan && (newOp & KikimrModifyOps())) {
                 TString message = TStringBuilder() << "Operation '" << newOp
                     << "' can't be performed in scan query";
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                 return {false, issues};
             }
@@ -393,7 +379,7 @@ public:
                     message = TStringBuilder() << message
                         << " Use COMMIT statement to indicate end of transaction between scheme and data operations.";
                 }
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
+
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_MIXED_SCHEME_DATA_TX, message));
                 return {false, issues};
             }
@@ -401,20 +387,18 @@ public:
             if (Readonly && (newOp & KikimrModifyOps())) {
                 TString message = TStringBuilder() << "Operation '" << newOp
                     << "' can't be performed in read only transaction";
-                const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                 issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                 return {false, issues};
             }
 
             auto& currentOps = TableOperations[table];
-            const bool currentModify = currentOps & KikimrModifyOps();
+            bool currentModify = currentOps & KikimrModifyOps();
             if (currentModify) {
                 if (KikimrReadOps() & newOp) {
                     if (!EnableImmediateEffects) {
                         TString message = TStringBuilder() << "Data modifications previously made to table '" << table
                             << "' in current transaction won't be seen by operation: '"
                             << newOp << "'";
-                        const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                         auto newIssue = AddDmlIssue(YqlIssue(pos, TIssuesIds::KIKIMR_READ_MODIFIED_TABLE, message));
                         issues.AddIssue(newIssue);
                         return {false, issues};
@@ -423,11 +407,10 @@ public:
                     HasUncommittedChangesRead = true;
                 }
 
-                if ((*info)->GetHasIndexTables()) {
+                if (info->GetHasIndexTables()) {
                     if (!EnableImmediateEffects) {
                         TString message = TStringBuilder()
                             << "Multiple modification of table with secondary indexes is not supported yet";
-                        const TPosition pos(op.GetPosition().GetColumn(), op.GetPosition().GetRow());
                         issues.AddIssue(YqlIssue(pos, TIssuesIds::KIKIMR_BAD_OPERATION, message));
                         return {false, issues};
                     }
@@ -445,9 +428,9 @@ public:
     virtual ~TKikimrTransactionContextBase() = default;
 
 public:
+    THashMap<TString, TYdbOperations> TableOperations;
     bool HasUncommittedChangesRead = false;
     const bool EnableImmediateEffects;
-    THashMap<TString, TYdbOperations> TableOperations;
     THashMap<TKikimrPathId, TString> TableByIdMap;
     TMaybe<NKikimrKqp::EIsolationLevel> EffectiveIsolationLevel;
     THashMap<TString, TString> TempTables;
