+++ ydb/core/kqp/session_actor/kqp_session_actor.cpp
@@ -2,7 +2,6 @@
 #include "kqp_tx.h"
 #include "kqp_worker_common.h"
 #include "kqp_query_state.h"
-#include "kqp_query_stats.h"
 
 #include <ydb/core/kqp/common/kqp.h>
 #include <ydb/core/kqp/common/kqp_lwtrace_probes.h>
@@ -633,7 +632,7 @@ public:
         }
 
         Counters->ReportTxCreated(Settings.DbCounters);
-        Counters->ReportBeginTransaction(Settings.DbCounters, Transactions.EvictedTx, Transactions.Size(), Transactions.ToBeAbortedSize()); // TODO: check 
+        Counters->ReportBeginTransaction(Settings.DbCounters, Transactions.EvictedTx, Transactions.Size(), Transactions.ToBeAbortedSize());
     }
 
     bool PrepareQueryTransaction() {
@@ -719,15 +718,16 @@ public:
         }
 
         try {
-            const auto& parameters = QueryState->GetYdbParameters();
-            QueryState->QueryData->ParseParameters(parameters);
-            if (QueryState->CompileResult && QueryState->CompileResult->Ast && QueryState->CompileResult->Ast->PgAutoParamValues) {
-                for(const auto& [name, param] : *QueryState->CompileResult->Ast->PgAutoParamValues) {
-                    if (!parameters.contains(name)) {
-                        QueryState->QueryData->AddTypedValueParam(name, param);
+            auto parameters = QueryState->GetYdbParameters();
+            if (QueryState->CompileResult && QueryState->CompileResult->Ast) {
+                auto& params = QueryState->CompileResult->Ast->PgAutoParamValues;
+                if (params) {
+                    for(const auto& [name, param] : *params) {
+                        parameters.insert({name, param});
                     }
                 }
             }
+            QueryState->QueryData->ParseParameters(parameters);
         } catch(const yexception& ex) {
             ythrow TRequestFail(Ydb::StatusIds::BAD_REQUEST) << ex.what();
         }
@@ -739,7 +739,7 @@ public:
         QueryResponse = std::make_unique<TEvKqp::TEvQueryResponse>();
         FillCompileStatus(QueryState->CompileResult, QueryResponse->Record);
 
-        auto ru = NRuCalc::CpuTimeToUnit(TDuration::MicroSeconds(QueryState->CompileStats.CpuTimeUs));
+        auto ru = NRuCalc::CpuTimeToUnit(TDuration::MicroSeconds(QueryState->CompileStats.GetCpuTimeUs()));
         auto& record = QueryResponse->Record.GetRef();
         record.SetConsumedRu(ru);
 
@@ -1157,7 +1157,7 @@ public:
                     NKqpProto::TKqpStatsQuery& stats = *ev->Get()->Record.MutableQueryStats();
                     NKqpProto::TKqpStatsQuery executionStats;
                     executionStats.Swap(&stats);
-                    stats = QueryState->QueryStats.ToProto();
+                    stats = QueryState->Stats;
                     stats.MutableExecutions()->MergeFrom(executionStats.GetExecutions());
                     ev->Get()->Record.SetQueryPlan(SerializeAnalyzePlan(stats));
                 }
@@ -1283,8 +1283,8 @@ public:
         }
 
         if (executerResults.HasStats()) {
-            QueryState->QueryStats.Executions.emplace_back();
-            QueryState->QueryStats.Executions.back().Swap(executerResults.MutableStats());
+            auto* exec = QueryState->Stats.AddExecutions();
+            exec->Swap(executerResults.MutableStats());
         }
 
         if (!response->GetIssues().empty()){
@@ -1323,7 +1323,7 @@ public:
         }
     }
 
-    void CollectSystemViewQueryStats(const TKqpQueryStats* stats, TDuration queryDuration,
+    void CollectSystemViewQueryStats(const NKqpProto::TKqpStatsQuery* stats, TDuration queryDuration,
         const TString& database, ui64 requestUnits)
     {
         auto type = QueryState->GetType();
@@ -1339,7 +1339,7 @@ public:
                 TString text = QueryState->ExtractQueryText();
                 if (IsQueryAllowedToLog(text)) {
                     auto userSID = QueryState->UserToken->GetUserSID();
-                    CollectQueryStats(TlsActivationContext->AsActorContext(), stats, queryDuration, text,
+                    NSysView::CollectQueryStats(TlsActivationContext->AsActorContext(), stats, queryDuration, text,
                         userSID, QueryState->ParametersSize, database, type, requestUnits);
                 }
                 break;
@@ -1351,19 +1351,16 @@ public:
 
     void FillSystemViewQueryStats(NKikimrKqp::TEvQueryResponse* record) {
         YQL_ENSURE(QueryState);
-        auto* stats = &QueryState->QueryStats;
+        auto* stats = &QueryState->Stats;
 
-        stats->DurationUs = ((TInstant::Now() - QueryState->StartTime).MicroSeconds());
-        stats->WorkerCpuTimeUs = (QueryState->GetCpuTime().MicroSeconds());
+        stats->SetDurationUs((TInstant::Now() - QueryState->StartTime).MicroSeconds());
+        stats->SetWorkerCpuTimeUs(QueryState->GetCpuTime().MicroSeconds());
         if (QueryState->CompileResult) {
-            stats->Compilation.emplace();
-            stats->Compilation->FromCache = (QueryState->CompileStats.FromCache);
-            stats->Compilation->DurationUs = (QueryState->CompileStats.DurationUs);
-            stats->Compilation->CpuTimeUs = (QueryState->CompileStats.CpuTimeUs);
+            stats->MutableCompilation()->Swap(&QueryState->CompileStats);
         }
 
         if (IsExecuteAction(QueryState->GetAction())) {
-            auto ru = CalcRequestUnit(*stats);
+            auto ru = NRuCalc::CalcRequestUnit(*stats);
 
             if (record != nullptr) {
                 record->SetConsumedRu(ru);
@@ -1384,19 +1381,19 @@ public:
         auto requestInfo = TKqpRequestInfo(QueryState->UserRequestContext->TraceId, SessionId);
 
         if (IsExecuteAction(QueryState->GetAction())) {
-            auto queryDuration = TDuration::MicroSeconds(QueryState->QueryStats.DurationUs);
+            auto queryDuration = TDuration::MicroSeconds(QueryState->Stats.GetDurationUs());
             SlowLogQuery(TlsActivationContext->AsActorContext(), Config.Get(), requestInfo, queryDuration,
                 record->GetYdbStatus(), QueryState->UserToken, QueryState->ParametersSize, record,
                 [this]() { return this->QueryState->ExtractQueryText(); });
         }
 
+        auto* stats = &QueryState->Stats;
         if (QueryState->ReportStats()) {
-            auto stats = QueryState->QueryStats.ToProto();
             if (QueryState->GetStatsMode() >= Ydb::Table::QueryStatsCollection::STATS_COLLECTION_FULL) {
-                response->SetQueryPlan(SerializeAnalyzePlan(stats));
+                response->SetQueryPlan(SerializeAnalyzePlan(*stats));
                 response->SetQueryAst(QueryState->CompileResult->PreparedQuery->GetPhysicalQuery().GetQueryAst());
             }
-            response->MutableQueryStats()->Swap(&stats);
+            response->MutableQueryStats()->Swap(stats);
         }
     }
 
@@ -1436,7 +1433,7 @@ public:
             Counters->ReportQueryWithRangeScan(Settings.DbCounters);
         }
 
-        auto& stats = QueryState->QueryStats;
+        auto& stats = QueryState->Stats;
 
         ui32 affectedShardsCount = 0;
         ui64 readBytesCount = 0;
@@ -1452,9 +1449,9 @@ public:
         Counters->ReportQueryAffectedShards(Settings.DbCounters, affectedShardsCount);
         Counters->ReportQueryReadRows(Settings.DbCounters, readRowsCount);
         Counters->ReportQueryReadBytes(Settings.DbCounters, readBytesCount);
-        Counters->ReportQueryReadSets(Settings.DbCounters, stats.ReadSetsCount);
-        Counters->ReportQueryMaxShardReplySize(Settings.DbCounters, stats.MaxShardReplySize);
-        Counters->ReportQueryMaxShardProgramSize(Settings.DbCounters, stats.MaxShardProgramSize);
+        Counters->ReportQueryReadSets(Settings.DbCounters, stats.GetReadSetsCount());
+        Counters->ReportQueryMaxShardReplySize(Settings.DbCounters, stats.GetMaxShardReplySize());
+        Counters->ReportQueryMaxShardProgramSize(Settings.DbCounters, stats.GetMaxShardProgramSize());
     }
 
     void ReplySuccess() {
