+++ ydb/core/kqp/session_actor/kqp_tx.h
@@ -319,73 +319,42 @@ struct TTxId {
     }
 };
 
-}
-
-template<>
-struct THash<NKikimr::NKqp::TTxId> {
-    inline size_t operator()(const NKikimr::NKqp::TTxId& id) const noexcept {
-        return THash<NKikimr::TULID>()(id.Id);
-    }
-};
-
-namespace NKikimr::NKqp {
-    
 class TTransactionsCache {
-    size_t MaxActiveSize;
-    THashMap<TTxId, TIntrusivePtr<TKqpTransactionContext>, THash<NKikimr::NKqp::TTxId>> Active;
+    TLRUCache<TTxId, TIntrusivePtr<TKqpTransactionContext>> Active;
     std::deque<TIntrusivePtr<TKqpTransactionContext>> ToBeAborted;
-
-    auto FindOldestTransaction() {
-        if (Active.empty()) {
-            return std::end(Active);
-        }
-        auto oldest = std::begin(Active);
-        for (auto it = std::next(oldest); it != std::end(Active); ++it) {
-            if (oldest->second->LastAccessTime > it->second->LastAccessTime) {
-                oldest = it;
-            }
-        }
-        return oldest;
-    }
-
 public:
     ui64 EvictedTx = 0;
     TDuration IdleTimeout;
 
     TTransactionsCache(size_t size, TDuration idleTimeout)
-        : MaxActiveSize(size)
+        : Active(size)
         , IdleTimeout(idleTimeout)
-    {
-        Active.reserve(MaxActiveSize);
+    {}
+
+    size_t Size() {
+        return Active.Size();
     }
 
-    size_t Size() const {
-        return Active.size();
-    }
-
-    size_t MaxSize() const {
-        return MaxActiveSize;
+    size_t MaxSize() {
+        return Active.GetMaxSize();
     }
 
     TIntrusivePtr<TKqpTransactionContext> Find(const TTxId& id) {
-        auto it = Active.find(id);
-        if (it != std::end(Active)) {
-            it->second->Touch();
-            return it->second;
+        if (auto it = Active.Find(id); it != Active.End()) {
+            it.Value()->Touch();
+            return *it;
         } else {
-            return nullptr;
+            return {};
         }
     }
 
-    TIntrusivePtr<TKqpTransactionContext> ReleaseTransaction(const TTxId& id) {
-        const auto it = Active.find(id);
-        if (it != std::end(Active)) {
-            auto result = std::move(it->second);
-            Active.erase(it);
-            return result;
-        } else {
-            return nullptr;
+    TIntrusivePtr<TKqpTransactionContext> ReleaseTransaction(const TTxId& txId) {
+        if (auto it = Active.FindWithoutPromote(txId); it != Active.End()) {
+            auto ret = std::move(it.Value());
+            Active.Erase(it);
+            return ret;
         }
+        return {};
     }
 
     void AddToBeAborted(TIntrusivePtr<TKqpTransactionContext> ctx) {
@@ -393,20 +362,20 @@ public:
     }
 
     bool RemoveOldTransactions() {
-        if (Active.size() < MaxActiveSize) {
-            return true;
-        }
-
-        auto oldestIt = FindOldestTransaction();
-        auto currentIdle = TInstant::Now() - oldestIt->second->LastAccessTime;
-        if (currentIdle >= IdleTimeout) {
-            oldestIt->second->Invalidate();
-            ToBeAborted.emplace_back(std::move(oldestIt->second));
-            Active.erase(oldestIt);
-            ++EvictedTx;
+        if (Active.Size() < Active.GetMaxSize()) {
             return true;
         } else {
-            return false;
+            auto it = Active.FindOldest();
+            auto currentIdle = TInstant::Now() - it.Value()->LastAccessTime;
+            if (currentIdle >= IdleTimeout) {
+                it.Value()->Invalidate();
+                ToBeAborted.emplace_back(std::move(it.Value()));
+                Active.Erase(it);
+                ++EvictedTx;
+                return true;
+            } else {
+                return false;
+            }
         }
     }
 
@@ -414,15 +383,15 @@ public:
         if (!RemoveOldTransactions()) {
             return false;
         }
-        return Active.emplace(txId, txCtx).second;
+        return Active.Insert(std::make_pair(txId, txCtx));
     }
 
     void FinalCleanup() {
-        for (auto& item : Active) {
-            item.second->Invalidate();
-            ToBeAborted.emplace_back(std::move(item.second));
+        for (auto it = Active.Begin(); it != Active.End(); ++it) {
+            it.Value()->Invalidate();
+            ToBeAborted.emplace_back(std::move(it.Value()));
         }
-        Active.clear();
+        Active.Clear();
     }
 
     size_t ToBeAbortedSize() {
@@ -443,3 +412,10 @@ bool NeedSnapshot(const TKqpTransactionContext& txCtx, const NYql::TKikimrConfig
 bool HasOlapTableInTx(const NKqpProto::TKqpPhyQuery& physicalQuery);
 
 }  // namespace NKikimr::NKqp
+
+template<>
+struct THash<NKikimr::NKqp::TTxId> {
+    inline size_t operator()(const NKikimr::NKqp::TTxId& id) const noexcept {
+        return THash<NKikimr::TULID>()(id.Id);
+    }
+};
