+++ ydb/core/kqp/session_actor/kqp_worker_actor.cpp
@@ -1,13 +1,12 @@
 #include "kqp_worker_common.h"
-#include "kqp_query_stats.h"
 
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/counters.h>
 #include <ydb/core/base/cputime.h>
 #include <ydb/core/protos/kqp.pb.h>
 #include <ydb/core/engine/mkql_proto.h>
-#include <ydb/core/kqp/common/kqp_ru_calc.h>
 #include <ydb/core/kqp/common/kqp_timeouts.h>
+#include <ydb/core/kqp/common/kqp_ru_calc.h>
 #include <ydb/core/kqp/gateway/kqp_metadata_loader.h>
 #include <ydb/core/kqp/host/kqp_host.h>
 #include <ydb/core/sys_view/service/sysview_service.h>
@@ -915,7 +914,7 @@ private:
                 TString text = ExtractQueryText();
                 if (IsQueryAllowedToLog(text)) {
                     auto userSID = QueryState->RequestEv->GetUserToken()->GetUserSID();
-                    CollectQueryStats(ctx, stats, queryDuration, text,
+                    NSysView::CollectQueryStats(ctx, stats, queryDuration, text,
                         userSID, QueryState->RequestEv->GetParametersSize(), database, type, requestUnits);
                 }
                 break;
