+++ ydb/core/kqp/session_actor/ya.make
@@ -7,7 +7,6 @@ SRCS(
     kqp_worker_actor.cpp
     kqp_worker_common.cpp
     kqp_query_state.cpp
-    kqp_query_stats.cpp
     kqp_temp_tables_manager.cpp
 )
