+++ ydb/core/sys_view/service/sysview_service.cpp
@@ -23,6 +23,118 @@ using namespace NActors;
 namespace NKikimr {
 namespace NSysView {
 
+constexpr size_t QUERY_TEXT_LIMIT = 4096;
+
+void CollectQueryStats(const TActorContext& ctx, const NKqpProto::TKqpStatsQuery* queryStats,
+    TDuration queryDuration, const TString& queryText,
+    const TString& userSID, ui64 parametersSize, const TString& database,
+    const NKikimrKqp::EQueryType type, ui64 requestUnits)
+{
+    if (!AppData()->FeatureFlags.GetEnableSystemViews()) {
+        return;
+    }
+
+    auto collectEv = MakeHolder<NSysView::TEvSysView::TEvCollectQueryStats>();
+    collectEv->Database = database;
+
+    auto& stats = collectEv->QueryStats;
+    auto& dataStats = *stats.MutableStats();
+    auto& shardsCpuTime = *stats.MutableShardsCpuTimeUs();
+    auto& computeCpuTime = *stats.MutableComputeCpuTimeUs();
+
+    auto nodeId = ctx.SelfID.NodeId();
+    stats.SetNodeId(nodeId);
+
+    stats.SetEndTimeMs(TInstant::Now().MilliSeconds());
+    stats.SetDurationMs(queryDuration.MilliSeconds());
+
+    stats.SetQueryTextHash(MurmurHash<ui64>(queryText.data(), queryText.size()));
+    if (queryText.size() > QUERY_TEXT_LIMIT) {
+        auto limitedText = queryText.substr(0, QUERY_TEXT_LIMIT);
+        stats.SetQueryText(limitedText);
+    } else {
+        stats.SetQueryText(queryText);
+    }
+
+    if (userSID) {
+        stats.SetUserSID(userSID);
+    }
+    stats.SetParametersSize(parametersSize);
+
+    TString strType;
+    switch (type) {
+    case NKikimrKqp::QUERY_TYPE_SQL_DML:
+    case NKikimrKqp::QUERY_TYPE_PREPARED_DML:
+        strType = "data";
+        break;
+    case NKikimrKqp::QUERY_TYPE_SQL_SCAN:
+        strType = "scan";
+        break;
+    case NKikimrKqp::QUERY_TYPE_SQL_SCRIPT:
+    case NKikimrKqp::QUERY_TYPE_SQL_SCRIPT_STREAMING:
+        strType = "script";
+        break;
+    default:
+        break;
+    }
+    stats.SetType(strType);
+
+    if (!queryStats || queryStats->GetExecutions().empty()) {
+        ctx.Send(NSysView::MakeSysViewServiceID(nodeId), std::move(collectEv));
+        return;
+    }
+
+    shardsCpuTime.SetMin(Max<ui64>());
+    computeCpuTime.SetMin(Max<ui64>());
+
+    auto aggregate = [] (NKikimrSysView::TStatsAggr& to, const NYql::NDqProto::TDqStatsAggr& from) {
+        to.SetMin(std::min(to.GetMin(), from.GetMin()));
+        to.SetMax(std::max(to.GetMax(), from.GetMax()));
+        to.SetSum(to.GetSum() + from.GetSum());
+        to.SetCnt(to.GetCnt() + from.GetCnt());
+    };
+
+    for (const NYql::NDqProto::TDqExecutionStats& exec : queryStats->GetExecutions()) {
+        NKqpProto::TKqpExecutionExtraStats execExtra;
+        if (exec.HasExtra()) {
+            bool ok = exec.GetExtra().UnpackTo(&execExtra);
+            Y_UNUSED(ok);
+        }
+
+        dataStats.SetPartitionCount(dataStats.GetPartitionCount() + execExtra.GetAffectedShards());
+
+        for (auto& table : exec.GetTables()) {
+            dataStats.SetReadRows(dataStats.GetReadRows() + table.GetReadRows());
+            dataStats.SetReadBytes(dataStats.GetReadBytes() + table.GetReadBytes());
+            dataStats.SetUpdateRows(dataStats.GetUpdateRows() + table.GetWriteRows());
+            dataStats.SetUpdateBytes(dataStats.GetUpdateBytes() + table.GetWriteBytes());
+            dataStats.SetDeleteRows(dataStats.GetDeleteRows() + table.GetEraseRows());
+        }
+
+        aggregate(shardsCpuTime, execExtra.GetShardsCpuTimeUs());
+        aggregate(computeCpuTime, execExtra.GetComputeCpuTimeUs());
+    }
+
+    if (queryStats->HasCompilation()) {
+        auto& compileStats = queryStats->GetCompilation();
+        stats.SetFromQueryCache(compileStats.GetFromCache());
+        stats.SetCompileDurationMs(compileStats.GetDurationUs() / 1'000);
+        stats.SetCompileCpuTimeUs(compileStats.GetCpuTimeUs());
+    }
+
+    stats.SetProcessCpuTimeUs(queryStats->GetWorkerCpuTimeUs());
+    stats.SetTotalCpuTimeUs(
+        stats.GetProcessCpuTimeUs() +
+        stats.GetCompileCpuTimeUs() +
+        stats.GetShardsCpuTimeUs().GetSum() +
+        stats.GetComputeCpuTimeUs().GetSum()
+    );
+
+    stats.SetRequestUnits(requestUnits);
+
+    ctx.Send(NSysView::MakeSysViewServiceID(nodeId), std::move(collectEv));
+}
+
 static void CopyCounters(NKikimrSysView::TDbCounters* diff,
     const NKikimrSysView::TDbCounters& current)
 {
